requirements.txt
discord.py
asyncio
requests
openpyxl
bs4
datetime